rename server/preference
