+++ server/routes/preferences.js
