+++ server/server.js
@@ -12,7 +12,7 @@ global.XMLHttpRequest = XMLHttpRequest
 const auth = require('./routes/auth')
 const profiles = require('./routes/profiles')
 const posts = require('./routes/posts')
-const preferences = require('./routes/addPreferences')
+const preferences = require('./routes/preferences')
 
 // Run server
 const app = express()
